Adding the ability to specify an alias in the query method (this may need to be added in other places).
The method would use the model name as the alias. This actually seems like a good idea and I thought it would work.

Instead, when I passed it an existing query via the 3rd argument, the final query would select from the given table twice as it would see the two aliases (the one I created and the one that is the full model name) as two tables. So, the where clauses created by me and the where clause created by the method would not operate on the same table. The result was basically that the where clause created by the method was ignored (as it was being applied to a completely different "table")